# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: English Learning App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [text_test, audio, reading, video]
        - name: ageMin
          in: query
          schema:
            type: integer
        - name: ageMax
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"

    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create assignment (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update assignment (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentRequest"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete assignment (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, parent, admin]
        age:
          type: integer
          nullable: true
        totalPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - role
        - totalPoints
        - createdAt
    UserWithStats:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            totalTimeMinutes:
              type: integer
            completedAssignments:
              type: integer
            averageScore:
              type: number
              nullable: true
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
    Assignment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [text_test, audio, reading, video]
        ageMin:
          type: integer
        ageMax:
          type: integer
        points:
          type: integer
        mediaUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - type
        - ageMin
        - ageMax
        - points
        - createdAt
    AssignmentDetail:
      allOf:
        - $ref: "#/components/schemas/Assignment"
        - type: object
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
            content:
              type: string
              nullable: true
              description: Reading text content for reading assignments
    Question:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
          nullable: true
          description: Only returned for admin role
        orderIndex:
          type: integer
      required:
        - id
        - text
        - options
        - orderIndex
    CreateAssignmentRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [text_test, audio, reading, video]
        ageMin:
          type: integer
        ageMax:
          type: integer
        points:
          type: integer
        mediaUrl:
          type: string
        content:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestionRequest"
      required:
        - title
        - description
        - type
        - ageMin
        - ageMax
        - points
    CreateQuestionRequest:
      type: object
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        orderIndex:
          type: integer
      required:
        - text
        - options
        - correctAnswer
        - orderIndex
    UpdateAssignmentRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        points:
          type: integer
        mediaUrl:
          type: string
        content:
          type: string
...